Refactor: template -> 탭 메뉴별로 파일분리
