+++ src/view/View.js
@@ -1,9 +1,7 @@
 import { TAB_MENUS_TEXT } from '../utils/constants.js';
-import {
-  getProduceAddText,
-  getVendingMachineManageText,
-  getProductPurchaseText,
-} from '../utils/template.js';
+import { getProduceAddTemplate } from '../utils/template/productAddTemplate.js';
+import { getProductPurchaseTemplate } from '../utils/template/productPurchaseTemplate.js';
+import { getVendingMachineManageTemplate } from '../utils/template/vendingMachineManageTemplate.js';
 
 class View {
   constructor() {
@@ -26,15 +24,15 @@ class View {
   }
 
   showProductAddScreen(productAddMenu) {
-    this.$content.innerHTML = getProduceAddText(productAddMenu);
+    this.$content.innerHTML = getProduceAddTemplate(productAddMenu);
   }
 
   showVendingMachineManageScreen(vendingMachineManageMenu) {
-    this.$content.innerHTML = getVendingMachineManageText(vendingMachineManageMenu);
+    this.$content.innerHTML = getVendingMachineManageTemplate(vendingMachineManageMenu);
   }
 
   showProductPurchaseScreen(productAddMenu, productPurchaseMenu) {
-    this.$content.innerHTML = getProductPurchaseText(productAddMenu, productPurchaseMenu);
+    this.$content.innerHTML = getProductPurchaseTemplate(productAddMenu, productPurchaseMenu);
   }
 }
